// src/components/datepicker.component.d.ts
import { ElementRef, EventEmitter } from "@angular/core";
import { DateService } from './datepicker.service';
export declare class DatePickerComponent {
    DatepickerService: DateService;
    static config: any;
    date: Date;
    min: Date;
    max: Date;
    callback: EventEmitter<string | Date>;
    hClasses: any[];
    dClasses: any[];
    full: boolean;
    calendar: boolean;
    dayscroll: ElementRef;
    yearscroll: ElementRef;
    today: Date;
    selectedDate: Date;
    tempDate: Date;
    dateList: Date[];
    cols: number[];
    rows: number[];
    weekdays: string[];
    months: string[];
    active: boolean;
    type: 'date' | 'string' | 'year' | 'month' | 'calendar';
    mode: 'calendar' | undefined;
    constructor(DatepickerService: DateService);
    initialize(): void;
    openModal(): void;
    getDaysOfWeek(): string[];
    ngAfterViewChecked(): void;
    getMonths(): string[];
    getYears(): Date[];
    createDateList(selectedDate: Date): void;
    getDate(row: number, col: number): Date;
    isDefined(date: Date | string): boolean;
    isDisabled(date: Date): boolean;
    isActualDate(date: Date): boolean;
    isActualMonth(month: number): boolean;
    isActualYear(year: number): boolean;
    isSelectedDate(date: Date): boolean;
    isSelectedMonth(month: number): boolean;
    isSelectedYear(year: number): boolean;
    changeType(val: 'date' | 'string' | 'year' | 'month' | 'calendar'): void;
    showType(val: 'date' | 'string' | 'year' | 'month' | 'calendar'): boolean;
    selectDate(date: Date): void;
    selectMonth(month: number): void;
    selectYear(year: number): void;
    getSelectedWeekday(): string;
    getSelectedMonth(): string;
    getTempMonth(): string;
    getTempYear(): number;
    onCancel(e: Event): void;
    onDone(e: Event): void;
    selectMonthOrYear(): void;
    limitTo(arr: Array<string> | string, limit: number): Array<string> | string;
    getMonthRows(): {}[];
    nextMonth(): void;
    prevMonth(): void;
}